+++ hspec-core/src/Test/Hspec/Core/Runner.hs
@@ -78,7 +78,7 @@ applyDryRun c
   | otherwise = id
   where
     markSuccess :: Item () -> Item ()
-    markSuccess item = item {itemExample = evaluateExample Success}
+    markSuccess item = item {itemExample = safeEvaluateExample Success}
 
     removeCleanup :: SpecTree () -> SpecTree ()
     removeCleanup spec = case spec of