+++ CHANGES.markdown
@@ -6,6 +6,7 @@
   - Read command-line options from config files `~/.hspec` and `.hspec`
   - Rename `Fail` constructor of `Result`-type to `Failure`
   - Add `FailureReason` type
+  - Add `Test.Hspec.Core.Spec.safeEvaluateExample`
 
 ## Changes in 2.3.2
   - Compatibility with HUnit 1.5